metrics: move Metrics trait to util
